Move On & Off to Set Function; Update Switch & Dimmer
