+++ insteon_mqtt/device/Switch.py
@@ -54,103 +54,12 @@ class Switch(functions.State, functions.Set, functions.Scene, Base):
         # Remote (mqtt) commands mapped to methods calls.  Add to the base
         # class defined commands.
         self.cmd_map.update({
-            'on' : self.on,
-            'off' : self.off,
             'set_flags' : self.set_flags,
             })
 
         # Update the group map with the groups to be paired and the handler
         # for broadcast messages from this group
         self.group_map.update({0x01: self.handle_on_off})
-
-    #-----------------------------------------------------------------------
-    def on(self, group=0x01, level=None, mode=on_off.Mode.NORMAL, reason="",
-           transition=None, on_done=None):
-        """Turn the device on.
-
-        NOTE: This does NOT simulate a button press on the device - it just
-        changes the state of the device.  It will not trigger any responders
-        that are linked to this device.  To simulate a button press, call the
-        scene() method.
-
-        This will send the command to the device to update it's state.  When
-        we get an ACK of the result, we'll change our internal state and emit
-        the state changed signals.
-
-        Args:
-          group (int):  The group to send the command to.  For switches this
-                this must be 1.  Allowing a group here gives us a consistent
-                API to the on command across devices.
-          level (int):  If non-zero, turn the device on.  The API is an int
-                to keep a consistent API with other devices.
-          mode (on_off.Mode): The type of command to send (normal, fast, etc).
-          reason (str):  This is optional and is used to identify why the
-                 command was sent. It is passed through to the output signal
-                 when the state changes - nothing else is done with it.
-          on_done: Finished callback.  This is called when the command has
-                   completed.  Signature is: on_done(success, msg, data)
-        """
-        LOG.info("Switch %s cmd: on %s", self.addr, mode)
-        assert group == 0x01
-        assert isinstance(mode, on_off.Mode)
-
-        if transition or mode == on_off.Mode.RAMP:
-            LOG.error("Device %s does not support transition.", self.addr)
-            mode = on_off.Mode.NORMAL if mode == on_off.Mode.RAMP else mode
-
-        # Send the requested on code value.
-        cmd1 = on_off.Mode.encode(True, mode)
-        msg = Msg.OutStandard.direct(self.addr, cmd1, 0xff)
-
-        # Use the standard command handler which will notify us when the
-        # command is ACK'ed.
-        callback = functools.partial(self.handle_ack, reason=reason)
-        msg_handler = handler.StandardCmd(msg, callback, on_done)
-
-        self.send(msg, msg_handler)
-
-    #-----------------------------------------------------------------------
-    def off(self, group=0x01, mode=on_off.Mode.NORMAL, reason="",
-            transition=None, on_done=None):
-        """Turn the device off.
-
-        NOTE: This does NOT simulate a button press on the device - it just
-        changes the state of the device.  It will not trigger any responders
-        that are linked to this device.  To simulate a button press, call the
-        scene() method.
-
-        This will send the command to the device to update it's state.  When
-        we get an ACK of the result, we'll change our internal state and emit
-        the state changed signals.
-
-        Args:
-          group (int):  The group to send the command to.  For this device,
-                this must be 1.  Allowing a group here gives us a consistent
-                API to the on command across devices.
-          mode (on_off.Mode): The type of command to send (normal, fast, etc).
-          reason (str):  This is optional and is used to identify why the
-                 command was sent. It is passed through to the output signal
-                 when the state changes - nothing else is done with it.
-          on_done: Finished callback.  This is called when the command has
-                   completed.  Signature is: on_done(success, msg, data)
-        """
-        LOG.info("Switch %s cmd: off %s", self.addr, mode)
-        assert group == 0x01
-        assert isinstance(mode, on_off.Mode)
-
-        if transition or mode == on_off.Mode.RAMP:
-            LOG.error("Device %s does not support transition.", self.addr)
-            mode = on_off.Mode.NORMAL if mode == on_off.Mode.RAMP else mode
-
-        # Send an off or instant off command.
-        cmd1 = on_off.Mode.encode(False, mode)
-        msg = Msg.OutStandard.direct(self.addr, cmd1, 0x00)
-
-        # Use the standard command handler which will notify us when the
-        # command is ACK'ed.
-        callback = functools.partial(self.handle_ack, reason=reason)
-        msg_handler = handler.StandardCmd(msg, callback, on_done)
-        self.send(msg, msg_handler)
 
     #-----------------------------------------------------------------------
     def set_backlight(self, level, on_done=None):
@@ -324,34 +233,6 @@ class Switch(functions.State, functions.Set, functions.Scene, Base):
 
         # Current on/off level is stored in cmd2 so update our level.
         self._set_state(is_on=msg.cmd2 > 0x00, reason=on_off.REASON_REFRESH)
-
-    #-----------------------------------------------------------------------
-    def handle_ack(self, msg, on_done, reason=""):
-        """Callback for standard commanded messages.
-
-        This callback is run when we get a reply back from one of our
-        commands to the device.  If the command was ACK'ed, we know it worked
-        so we'll update the internal state of the device and emit the signals
-        to notify others of the state change.
-
-        Args:
-          msg (message.InpStandard):  The reply message from the device.
-              The on/off level will be in the cmd2 field.
-          on_done: Finished callback.  This is called when the command has
-                   completed.  Signature is: on_done(success, msg, data)
-          reason (str):  This is optional and is used to identify why the
-                 command was sent. It is passed through to the output signal
-                 when the state changes - nothing else is done with it.
-        """
-        # If this it the ACK we're expecting, update the internal state and
-        # emit our signals.
-        LOG.debug("Switch %s ACK: %s", self.addr, msg)
-
-        is_on, mode = on_off.Mode.decode(msg.cmd1)
-        reason = reason if reason else on_off.REASON_COMMAND
-        self._set_state(is_on=is_on, mode=mode, reason=reason)
-        on_done(True, "Switch state updated to on=%s" % self._is_on,
-                self._is_on)
 
     #-----------------------------------------------------------------------
     def handle_group_cmd(self, addr, msg):
